% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geospatial.R
\name{clean_addresses}
\alias{clean_addresses}
\title{Clean an Address Vector}
\usage{
clean_addresses(addresses)
}
\arguments{
\item{addresses}{A character vector of addresses}
}
\value{
A character vector with title case capitalization, with corrected single letter and direction capitalization
}
\description{
This function cleans addresses by making them more readable, primarily correcting case.
}
