% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidycensus_supplement.R
\name{create_var_list}
\alias{create_var_list}
\title{Create ACS Variable Vector}
\usage{
create_var_list(table, var_nums)
}
\arguments{
\item{table}{The name of an ACS table, e.g. B03002}

\item{var_nums}{The variable numbers to pull from an ACS table}
}
\description{
This function returns a character vector of ACS variables, pasting together a table name that is entered and a vector of variable numbers
}
\examples{
# Create variables for table B02001 for variable numbers 1-8
# These are the top-level variables in the ACS race table
create_var_list("B02001", 1:8)
}
