% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ia_bg}
\alias{ia_bg}
\title{Iowa Census Block Group Data}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 2703 rows and 21 columns.
}
\source{
U.S. Census Bureau TIGER/LINE
}
\usage{
ia_bg
}
\description{
A data frame of Iowa block groups. Data from tigris, with columns added for county, tract, place, and zip code information. Corresponding primary places and ZIP codes are determined by the place or ZIP code that takes up the largest amount of the block group's area.
}
\keyword{datasets}
