% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ia_places}
\alias{ia_places}
\title{Iowa Place Data}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 1028 rows and 18 columns.
}
\source{
U.S. Census Bureau TIGER/LINE
}
\usage{
ia_places
}
\description{
A data frame of Iowa places. Data from tigris, with a column added for county, determined by which county the place centroid is located in.
}
\keyword{datasets}
