% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ia_tracts}
\alias{ia_tracts}
\title{Iowa Census Tract Data}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 896 rows and 15 columns.
}
\source{
U.S. Census Bureau TIGER/LINE
}
\usage{
ia_tracts
}
\description{
A data frame of Iowa places. Data from tigris, with a column added for a list of places that intersect with each Census tract.
}
\keyword{datasets}
