% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ia_zctas}
\alias{ia_zctas}
\title{Iowa ZCTA Data}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 935 rows and 14 columns.
}
\source{
U.S. Census Bureau TIGER/LINE
}
\usage{
ia_zctas
}
\description{
A data frame of Iowa zip code tabulation areas. Data from tigris, with columns added for place and county lists. The lists are determined by places and counties in which the ZCTA intersects.
}
\keyword{datasets}
