% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psc_theme.R
\name{psc_colors}
\alias{psc_colors}
\title{PSC color scheme}
\usage{
psc_colors(num)
}
\arguments{
\item{num}{A number from 1 to 10}
}
\value{
A character vector of PSC color hex codes
}
\description{
This function returns a specified number of colors (up to 10) in the PSC color scheme
}
