% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geospatial.R
\name{st_view}
\alias{st_view}
\title{View an sf data frame}
\usage{
st_view(sf_data)
}
\arguments{
\item{sf_data}{An object with class 'sf' and class 'data.frame'}
}
\description{
This function removes the geometry from an sf data frame for purposes of viewing the data in RStudio.
This can be useful because sf geometry can significantly slow down RStudio's View() function
}
